ccct/ccct.py: Avoid _parse_args() failure
If _parse_args() fails in a way that prevents the args variable from
being populated, the _load_from_config() function needs to detect this
and raise an exception. There does not seem to be any way for this to
happen outside of contrived testing scenarios, but it cannot be ruled
out.

Signed-off-by: Chuck Wolber <chuckwolber@gmail.com>