ccct/ccct.py: Support unit testing
Add return values and additional arguments to support unit testing.
This enables pass/fail detection when functions return and allows for
alternate schema and config files to be passed in to test for error
conditions.

Made _resolve_config() the owner of the configuration control flow to
avoid situations where _load_from_config() does not have an args
object.

Added the complete configuration set to _resolve_config() to ensure
a full final validation happens in one place. This allows testing
with zero args and a config file set to 'None'.

Signed-off-by: Chuck Wolber <chuckwolber@gmail.com>